change offset to 16 bit

+++ src/gfx.c
@@ -12,9 +12,9 @@
 #include "gfx.h"
 #include "ssd1306.h"
 
-int gfx_waveform_64_32(uint8_t x, uint8_t y, uint8_t color, const uint16_t *wform, uint8_t offset, uint8_t rshift)
+int gfx_waveform_64_32(uint8_t x, uint8_t y, uint8_t color, const uint16_t *wform, uint16_t offset, uint8_t rshift)
 {
-	uint8_t i,o;
+	uint16_t i,o;
 
 	// Make sure the offset is in range
 	if (offset >= 64)
