+++ src/testers/gfx_tester.c
@@ -45,7 +45,7 @@ static const uint16_t gfx_tester_expdecay_wave[64] = {
 	0x01f, 0x01d, 0x01b, 0x01a, 0x018, 0x017, 0x015, 0x014
 };
 
-static int gfx_tester_scope(const uint16_t *waveform, uint8_t offset)
+static int gfx_tester_scope(const uint16_t *waveform, uint16_t offset)
 {
 	gfx_graticule_cfg_t grcfg =
 	{
@@ -231,7 +231,7 @@ static int gfx_tester_voltmeter(void)
 
 int gfx_tester_run(void)
 {
-	uint8_t i;
+	uint16_t i;
 
 	// Render a scrolling sine wave
 	for (i=0; i<64; i++)
